+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT UP
 
 - Effect: monsters with Uncanny Dodge will get the Uncanny Dodge effect on the actor item.
+- Monster effect: Bard Taunt. (thanks @Moto Moto#0295)
 
 
 # 3.4.30